requirements.txt
requests
PyYAML
pydantic
privatbank-api-client
python-telegram-bot